--- src/models/Animals/Fishes.java
@@ -1,5 +1,0 @@
-package models.Animals;
-
-public class Fishes {
-    
-}
